+++ Flank/src/main/java/com/walmart/otto/tools/GcloudTool.java
@@ -14,6 +14,7 @@ import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 public class GcloudTool extends Tool {
+
   private boolean printTests = true;
   private boolean testFailed = false;
   private int latestExecutionTime;
@@ -37,6 +38,8 @@ public class GcloudTool extends Tool {
           "instrumentation",
           orchestratorFlag(),
           autoGoogleLoginFlag(),
+          recordVideoFlag(),
+          performanceMetricsFlag(),
           "--app",
           bucket + getSimpleName(getAppAPK()),
           "--test",
@@ -84,6 +87,20 @@ public class GcloudTool extends Tool {
   private String autoGoogleLoginFlag() {
     if (getConfigurator().disableAutoGoogleLogin()) {
       return "--no-auto-google-login";
+    }
+    return "";
+  }
+
+  private String recordVideoFlag() {
+    if (!getConfigurator().recordVideo()) {
+      return "--no-record-video";
+    }
+    return "";
+  }
+
+  private String performanceMetricsFlag() {
+    if (!getConfigurator().recordPerformanceMetrics()) {
+      return "--no-performance-metrics";
     }
     return "";
   }