+++ release_notes.md
@@ -8,3 +8,4 @@ Release notes for flank
 
 - [Add IntelliJ Idea Code Formatting settings for Java Flank - **sridoodla**](https://github.com/TestArmada/flank/pull/147)
 - [Add ability to disable google login - **runningcode**](https://github.com/TestArmada/flank/pull/146)
+- [Add the ability to disable video recording and performance metrics - **sridoodla**](https://github.com/TestArmada/flank/pull/149)